backend/requirements.txt
# langchain
# langchain-community
# langchain-ollama
# ollama
# python-dotenv
# youtube-transcript-api



fastapi
uvicorn
python-dotenv
youtube-transcript-api
langchain
langchain_ollama
